refactor, make use of try catch
