+++ SocketIO-iOSTests/SocketParserTest.swift
@@ -17,8 +17,7 @@ class SocketParserTest: XCTestCase {
         "51-/swift,[\"testMultipleItemsWithBufferEmitReturn\",[1,2],{\"test\":\"bob\"},25,\"polo\",{\"_placeholder\":true,\"num\":0}]": ("/swift", ["testMultipleItemsWithBufferEmitReturn", [1, 2], ["test": "bob"], 25, "polo", "~~0"], [], -1),
         "3/swift,0[[\"test3\",\"test4\"]]": ("/swift", [["test3", "test4"]], [], 0),
         "61-/swift,19[[1,2],{\"test\":\"bob\"},25,\"polo\",{\"_placeholder\":true,\"num\":0}]": ("/swift", [ [1, 2], ["test": "bob"], 25, "polo", "~~0"], [], 19),
-        "4/swift,": ("/swift", [], [], -1),
-        "10": ("/swift", [], [], -1)]
+        "4/swift,": ("/swift", [], [], -1)]
     
     func testDisconnect() {
         let message = "1"
@@ -57,8 +56,12 @@ class SocketParserTest: XCTestCase {
     
     func testInvalidInput() {
         let message = "8"
-        let packet = SocketParser.parseString(message)
-        XCTAssertNil(packet)
+        do {
+            try SocketParser.parseString(message)
+            XCTFail("Should throw exeption")
+        } catch {
+            
+        }
     }
     
     func testGenericParser() {
@@ -66,32 +69,27 @@ class SocketParserTest: XCTestCase {
         XCTAssertEqual(parser.read(1), "6")
         XCTAssertEqual(parser.currentCharacter, "1")
         XCTAssertEqual(parser.readUntilStringOccurence("-"), "1")
-       XCTAssertEqual(parser.currentCharacter, "-")
+        XCTAssertEqual(parser.currentCharacter, "-")
         
     }
     
     func validateParseResult(message:String) {
         let validValues = SocketParserTest.packetTypes[message]!
-        let packet = SocketParser.parseString(message)
+        let packet = try! SocketParser.parseString(message)
         let type = message.substringWithRange(Range<String.Index>(start: message.startIndex, end: message.startIndex.advancedBy(1)))
-        if let packet = packet {
-            XCTAssertEqual(packet.type, SocketPacket.PacketType(str:type)!)
-            XCTAssertEqual(packet.nsp, validValues.0)
-            XCTAssertTrue((packet.data as NSArray).isEqualToArray(validValues.1))
-            XCTAssertTrue((packet.binary as NSArray).isEqualToArray(validValues.2))
-            XCTAssertEqual(packet.id, validValues.3)
-        }else {
-            XCTFail()
-        }
+        XCTAssertEqual(packet.type, SocketPacket.PacketType(str:type)!)
+        XCTAssertEqual(packet.nsp, validValues.0)
+        XCTAssertTrue((packet.data as NSArray).isEqualToArray(validValues.1))
+        XCTAssertTrue((packet.binary as NSArray).isEqualToArray(validValues.2))
+        XCTAssertEqual(packet.id, validValues.3)
     }
     
     func testParsePerformance() {
         let keys = Array(SocketParserTest.packetTypes.keys)
         measureBlock({
             for item in keys.enumerate() {
-                SocketParser.parseString(item.element)
+                try! SocketParser.parseString(item.element)
             }
         })
-        
     }
 }
